Few cosmetic changes and fixed issues with JDBCReservationDAO

Approved-by: Jim Hofer <jimpgh2014@gmail.com>

+++ module-2/09_Capstone/java/src/main/java/com/techelevator/models/ReservationDAO.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface ReservationDAO {
 	
-	public List<Site> getAvailableReservations(long campgroundID, String startDate, String endDate, Integer personCapacity, Boolean needsWheelchairAccess, Integer rvLengthRequired, Boolean needsUtility);
+	public List<Site> getAvailableReservations(long campgroundID, String startDate, String endDate, Long personCapacity, Boolean needsWheelchairAccess, Integer rvLengthRequired, Boolean needsUtility);
 	public List<Site> getAvailableReservations(long campgroundID, String startDate, String endDate);
 	
 	public long makeReservation(long siteID, String name, String startDate, String endDate);
